refactor: split oversized cross_format and cleanup/apply tests.rs files (#2050)
